src/django_project/users_app/models.py
from django.db import models
from django.contrib.auth.models import User
from PIL import Image

class Profile(models.Model):
    user = models.OneToOneField(User, on_delete=models.CASCADE, verbose_name='Пользователь')
    bio = models.TextField(verbose_name='Биография')
    social_link = models.URLField(max_length=200, blank=True, null=True)
    avatar = models.ImageField(upload_to='avatar/', blank=True, null=True, verbose_name='Аватар')

    class Meta:
        verbose_name='Профиль'
        verbose_name_plural = 'Профили'

    def __str__(self):
        return self.user.username

    def save(self, *args, **kwargs):
        super().save(*args, **kwargs)
        if self.avatar:
            img = Image.open(self.avatar.path)
            if img.height > 300 or img.width > 300:
                img.thumbnail((300, 300))
                img.save(self.avatar.path)
