pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "anastasia-anufrikova", email = "nastyaanufrikova1996@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-jazzmin>=3.0.2",
    "pillow>=12.1.1",
    "pre-commit>=4.5.0",
]

[project.scripts]
django-project = "django_project:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
